# pkg/app/response.go
package app

import (
	"github.com/gin-gonic/gin"
	"github.com/wudidayan/go-blog/pkg/e"
)

type Gin struct {
	C *gin.Context
}

type Response struct {
	Code int         `json:"code"`
	Msg  string      `json:"msg"`
	Data interface{} `json:"data"`
}

func (g *Gin) Response(httpCode, errCode int, data interface{}) {
	if data == nil {
		g.C.JSON(httpCode, Response{
			Code: errCode,
			Msg:  e.GetMsg(errCode),
			Data: make(map[string]interface{}),
		})
	} else {
		g.C.JSON(httpCode, Response{
			Code: errCode,
			Msg:  e.GetMsg(errCode),
			Data: data,
		})
	}

	return
}